+++ src/bask_task.c
@@ -170,11 +170,11 @@ static int task_upgrade (struct bask_task* task)
 }
 
 /*
-	Function: task_insert (struct bask_task** first, unsigned int n, unsigned int tid, int tactive, short tpriority, int tstate, char* tadded, char* tdue, char* tfinished, char* tproject, char* tdescription);
+	Function: task_insert (struct bask_task** first, unsigned int n, unsigned int tid, short tactive, short tpriority, int tstate, char* tadded, char* tdue, char* tfinished, char* tproject, char* tdescription);
 	Description: Inserts a new task with data.
 	InitVersion: 0.0.1
 */
-int task_insert (struct bask_task** first, unsigned int n, unsigned int tid, int tactive, short tpriority, int tstate, char* tadded, char* tdue, char* tfinished, char* tproject, char* tdescription)
+int task_insert (struct bask_task** first, unsigned int n, unsigned int tid, short tactive, short tpriority, int tstate, char* tadded, char* tdue, char* tfinished, char* tproject, char* tdescription)
 {
 	struct bask_task *newobj = malloc (sizeof (struct bask_task)), *preobj;
 	
@@ -280,11 +280,11 @@ int task_remove (struct bask_task** first, unsigned int id)
 }
 
 /*
-	Function: task_modificate (struct bask_task** first, unsigned int id, int active, int state, short priority, char* added, char* due, char* finished, char* project, char* description);
+	Function: task_modificate (struct bask_task** first, unsigned int id, short active, int state, short priority, char* added, char* due, char* finished, char* project, char* description);
 	Description: Modificates a task!
 	InitVersion: 0.0.1
 */
-int task_modificate (struct bask_task** first, unsigned int id, int active, int state, short priority, char* added, char* due, char* finished, char* project, char* description)
+int task_modificate (struct bask_task** first, unsigned int id, short active, int state, short priority, char* added, char* due, char* finished, char* project, char* description)
 {
 	struct bask_task* ptr = *first;
 	
@@ -482,17 +482,17 @@ int task_search (bask_core* tcore, struct bask_task** first, struct bask_task** 
    |--------------------------------------------| */
 
 /*
-	Function: task_create_cmd (bask_core* tcore, struct bask_task** first, int priority, char* project, char* description);
+	Function: task_create_cmd (bask_core* tcore, struct bask_task** first, short priority, char* project, char* description);
 	Description: Cmd handle for creating a task!
 	InitVersion: 0.0.1
 */
-void task_create_cmd (bask_core* tcore, struct bask_task** first, int priority, char* project, char* description)
+void task_create_cmd (bask_core* tcore, struct bask_task** first, short priority, char* project, char* description)
 {
 	/* TODO: Move the range check of the priority variable into the task_check_input function. */
 	if (task_check_input (tcore, "", "", "", project, description, 1) == 0 &&
-	    task_check_input_nbrs (0, (short) priority, 1, 1) == 0)
+	    task_check_input_nbrs (0, priority, 1, 1) == 0)
 	{
-		task_create (tcore, first, (short) priority, project, description);
+		task_create (tcore, first, priority, project, description);
 		printf ("Created task %i.\n", tcore->baskbin_uid);
 		export_baskbin (tcore, first, tcore->path_baskbin);
 	}
@@ -511,17 +511,17 @@ void task_remove_cmd (bask_core* tcore, struct bask_task** first, unsigned int i
 }
 
 /*
-	Function: task_modificate_cmd (bask_core* tcore, struct bask_task** first, unsigned int id, int active, int state, int priority, char* added, char* due, char* finished, char* project, char* description);
+	Function: task_modificate_cmd (bask_core* tcore, struct bask_task** first, unsigned int id, short active, int state, short priority, char* added, char* due, char* finished, char* project, char* description);
 	Description: Cmd handle for modifying a task!
 	InitVersion: 0.0.1
 */
-void task_modificate_cmd (bask_core* tcore, struct bask_task** first, unsigned int id, int active, int state, int priority, char* added, char* due, char* finished, char* project, char* description)
+void task_modificate_cmd (bask_core* tcore, struct bask_task** first, unsigned int id, short active, int state, short priority, char* added, char* due, char* finished, char* project, char* description)
 {
 	/* TODO: Move the range check of the priority variable into the task_check_input function. */
 	if (task_check_input (tcore, added, due, finished, project, description, 1) == 0 &&
-	    task_check_input_nbrs (id, (priority == -1 ? 0 : (short) priority), (active == -1 ? 0 : (short) active), 1) == 0)
+	    task_check_input_nbrs (id, (priority == -1 ? 0 : priority), (active == -1 ? 0 : active), 1) == 0)
 	{
-		task_modificate (first, id, active, state, (short) priority, added, due, finished, project, description);
+		task_modificate (first, id, active, state, priority, added, due, finished, project, description);
 		printf ("Modificated task %i.\n", id);
 		export_baskbin (tcore, first, tcore->path_baskbin);
 	}
