+++ src/bask_task.h
@@ -26,9 +26,9 @@ int task_check_input (bask_core*, char*, char*, char*, char*, char*, short);
 int task_get_priority (char*);
 
 void task_free_ll (struct bask_task**);
-int task_insert (struct bask_task**, unsigned int, unsigned int, int, short, int, char*, char*, char*, char*, char*);
+int task_insert (struct bask_task**, unsigned int, unsigned int, short, short, int, char*, char*, char*, char*, char*);
 int task_remove (struct bask_task**, unsigned int);
-int task_modificate (struct bask_task**, unsigned int, int, int, short, char*, char*, char*, char*, char*);
+int task_modificate (struct bask_task**, unsigned int, short, int, short, char*, char*, char*, char*, char*);
 
 int task_create (bask_core*, struct bask_task**, short, char*, char*);
 void task_deactivate (struct bask_task**, unsigned int);
@@ -37,9 +37,9 @@ int task_finish (struct bask_task**, unsigned int);
 
 int task_search (bask_core*, struct bask_task**, struct bask_task**, char*);
 
-void task_create_cmd (bask_core*, struct bask_task**, int, char*, char*);
+void task_create_cmd (bask_core*, struct bask_task**, short, char*, char*);
 void task_remove_cmd (bask_core*, struct bask_task**, unsigned int);
-void task_modificate_cmd (bask_core*, struct bask_task**, unsigned int, int, int, int, char*, char*, char*, char*, char*);
+void task_modificate_cmd (bask_core*, struct bask_task**, unsigned int, short, int, short, char*, char*, char*, char*, char*);
 void task_deactivate_cmd (bask_core*, struct bask_task**, unsigned int);
 void task_due_cmd (bask_core*, struct bask_task**, unsigned int, char*);
 void task_finish_cmd (bask_core*, struct bask_task**, unsigned int);
